feat: --device flag to specify which yubikey to use
When ykman detects multiple devices connected to the host it
refuses to operate unless the user specifies the yubikey serial
number using the `--device` flag.

This change implements the same (optional) `--device` flag in rofyk
so that if the user specify it is added to `ykman`'s args for
the underlying operation.